Add files via upload
Exploratory Data Analysis is done for bank data set (bank.csv). l1 and l2 are cleaned data of independent and dependent variables. the support vector machine(SVM) algorithm is used to predict the customer subscribe the term deposit or not. It gives the more accuracy than other classification algorithms. The random forest with 1000 trees nearly give the same accuracy , but the svm gives more than that and low computation time.